Allow disabling inputs in specific regions.
Replaces the older convention of setting the input minimum and maximum
to 0, which had the effect of disabling the slider in the ET-Engine
client.

Now, the input_values attribute on Area may have a "disabled: true"
attribute, and this will be included in user_values.json and
input_values.json so that client apps can disable the slider correctly.