Add autobalancing support to the v3 API.
Like v2, add an "autobalance=true" parameter to enable automatic
balancing of input groups.

Added ScenarioUpdatePresenter which receives a ScenarioUpdater and
outputs JSON for the API. The presenter, combined with the Scenario
Updater, asserts that the request is valid -- checking that the
inputs exist, values are within the permitted extrema, and that
gqueries exist -- before returning the scenario JSON and Gquery
results.

The new presenter is more strict than the previous implementation; any
invalid input value will abort updating the scenario.