Add INPUT_VALUE GQL function. Closes #515.
This function allows the GQL author to query the value of an input. If
the user has set a value, this will be returned. If the user has not set
an explicit value, but the input has been used in auto-balancing, the
balanced value is returned instead. If neither is present, the input's
default start value is returned.