Add Osmosis for balancing groups.
Osmosis is an extraction of the ETengine balance to plain Ruby (i.e.
without ETengine's input classes and caching). It was initially created
for use in the Optimizer, and is now added to the Engine.

A Balancer class is retained to support translating values into and back
from the data structures used by Osmosis.