Automatically calculate datasets using Atlas and Refinery
Production and Staging
----------------------

Rake tasks (and Capistrano hooks) have been added to run the Atlas and
Refinery calculations for each dataset when the application is deployed.
Loading a new ETSource commit (via the admin interface) will also cause
each dataset to be recalculated.

Development
-----------

In development, each region is instead "lazy-loaded" so that you don't
have to wait 5-10 seconds multiplied by the number of regions, but only
for the region you want to view. If you load a scenario for a new
region, the dataset calculations for that region will be calculated at
that time.

The "live-reloader" has been updated to expire the calculated datasets
whenever modifications to ETSource are detected. As a result, the
"etsource_cache_dataset" config option has been removed and is ALWAYS
enabled. This prevents multi-second load times for every ETSource page
in development; such long load times will happen only on the first
request.

I recommend you delete your config/config.yml and create a new one from
the sample.