Disable Memcached in development environment
With memcache, and the recent removal of the "do not cache dataset"
option, it was necessary to remember to flush memcache each time you
started the application, since an old dataset could still be in
Memcache.

Now we use the standard in-memory cache in development (which should be
fine since you only ever have a single Rails process in dev) which is
wiped out each time the application shuts down.