Change the cached Atlas dataset to use MessagePack
The use of MessagePack to store the calculated Atlas+Refinery data
replaces the previous use of YAML. MessagePack was chosen over YAML and
JSON because of it's vastly superior speed.

The performance of JSON is not far off MessagePack, so if the additional
depdendency is undesireable, swapping to JSON is not likely to pose a
problem.

                        user     system      total        real
    SAVE: YAML      0.890000   0.000000   0.890000 (  0.891219)
    READ: YAML      0.690000   0.000000   0.690000 (  0.685397)
    SAVE: JSON      0.060000   0.000000   0.060000 (  0.058253)
    READ: JSON      0.050000   0.000000   0.050000 (  0.047736)
    SAVE: MSGPACK   0.000000   0.000000   0.000000 (  0.008411)
    READ: MSGPACK   0.030000   0.010000   0.040000 (  0.035689)