Overhaul ETSource imports
Rewrites Etsource::Commit to scan the ETSource version prior to
importing to see what versions of Atlas and Refinery were used to run
the ETS tests.

Uses the Git library as much as possible for the git-archive process,
and uses a new convention of storing the export in tmp/etsource,
removing the old ETS "latest_sha" file, instead opting to place a
REVISION file in the ETS export.