New ETSource commit listing page
Mimicks the commit log on GitHub. Shows when an ETSource is incompatible
with the current version of ETEngine (because it specifies versions of
Atlas or Refinery which are newer than the ones which are loaded).

Shows a warning - but allows the user to continue - if the ETS commit
was tested with an older versions of Atlas or Refinery than are loaded.