Add "primary_demand_of_carriers"
This altered version of "primary_demand_of_carrier" can accept multiple
carrier keys. It replaces the convention of calling "of_carrier"
multiple times and summing the result, which would cause energy to be
included multiple times if the carriers were related.

For example, calculating primary demand of network gas and natural gas
would cause the natural gas to be counted twice (once as part of the
network gas supply, and once on its own).