Swap AuthLogic for Devise
Implements a password migration from AuthLogic's 20-stretch SHA512 to
BCrypt. The old SHAs are hashed with BCrypt, and the first time a user
logs in we'll 20-stretch SHA512 their plain-text password, and check it
against the BCrypted version. If valid, we'll re-BCrypt the plain-text
password and store it in Devise's "envrypted_password" column, and
remove the AuthLogic-related data.

See http://blog.jgc.org/2012/06/one-way-to-fix-your-rubbish-password.html