Fix dependent supply when conversion is not 1.0
Previously it was assumed that when calculating dependent supply, any
node which ended the traversal through the graph would contribute all of
its demand to the calculation. This was wrong, since there may be
losses, or conversions to multiple output carriers. The fixed method now
accounts for the output conversion, instead of assuming 1.0

Closes #666