Add GraphApi#loss_of_load_probability
... and the "demand_curve" helper methods. The latter returns the
electricity load on the graph for each hour in the year, while the
loss of load probability returns what proportion of those hours
exceed a given production capacity limit.

Also added a "total_demand_for_electricity" method which sums the three
electricity demands in GraphApi, and simplified the Merit Order injector
with this.

Ref #669