Slot calculation should consider reverse shares
When a slot has a mixture of "forward" and "reversed" share links, the
reversed shares are not considered when calculating the demand of the
slot due to them being considered "active" links.

Therefore, slots with such a mixture need to have their "external_value"
altered to account for the as-yet-uncalculated reverse shares.

Ref #689