Support for scaling areas
Adds support for reducing the size of an area proportionally, according
to one area attribute. For example, the Netherlands has 7.3M residences;
you may wish to scale down the area to only 50,000 residences, to
simulate a mid-sized city.

All attributes within the area are reduced in proportion to the main
attribute, but edge shares and slot conversions are preserved. Input
min, max, and default values are also adjusted as appropriate.

Ref #712