Add class-based plugins for the graph
The MeritOrderInjector has been removed, and replaced with two plugins:
SimpleMeritOrder and MeritOrder. The former provides useful
merit-related helpers without requiring a full calculation of the order
(such as is always required on the present graph, or on the future graph
with the user has disabled the merit order calculation).

Other graph plugins, such as FCE, resettable slots, etc, have been
rewritten to be class-based.

A new Lifecycle class handles these plugins, and triggering their
callbacks at various points in the graph calculation.