Ignore NoU=nil and recompute the value
With the addition of a number_of_units attribute in Atlas, most
converters will now have an NoU set to nil.

DatasetAttributes#fetch considers a `nil` to be a cache hit (the key is
present) and returns nil as the value, instead of computing the block
given.

NoU set to nil will now be ignored, and the block executed to compute
the correct value.

Closes #813.