Add PowerToHeat adapter to merit order
Supports power-to-heat technologies in the merit order calculation.
Takes a demand profile, and a total demand from another converter, and
creates a storage technology which does not emit any energy back into
the merit order (it has been changed to heat).

The demand curve is subtracted from the contents of the buffer in each
step.

Ref quintel/merit#132
Ref quintel/merit#131