Add converter API which inherits NoU from a parent
Converters which belong to the "inheritable_nou" group will inherit
their number_of_units from their parent (left) node. A node which acts
this way is permitted to have only one parent (via a single output
link).

In order to determine the NoU, the value of the parent is taken and then
adjusted according to the slot conversion of the parent node and the
share of the connecting edge.

For example, a parent and child node connected via an electricity edge
of share 0.5 and with an electricity input conversion of 0.5, where the
parent has 100 units, will assign 25 units to the child.

Ref quintel/etmodel#2360