Remember number_of_units in Fever adapters
Prevents the number of units evalulating to zero for technologies
connected to useful demand with a flexible edge. The adapters would call
Converter#number_of_units during inject!, which is before the graph has
been calculated (for the second time) and the flexible edge share was
wrongly set to nil.

Closes #944