Fever plugin assigns only production to nodes
Fever participants which use the Storage producer type would use the
sum of the load curve to determine the demand of the node. This is
incorrect, since the load curve contains negatives to indicate net
charging and positives for net discharge. Summing this curve would
therefore result in a value close to zero.

This commit changes the adapter to ignore the negative loads and
herefore use the heat produced as the node demand.

Closes quintel/etmodel#2477