Merit::Util.add_curves works with any number of curves
Defines curve adders when given ten and twenty curves, and a general
"add_many" which provides an optimised way to add a number of curves for
which there is no specific adder.

add_many partitions the collection of curves into groups of 20, 10, or 5
curves and then delegates to the optimised adder methods.