Fix share of flexible links when share is manually set
If a user update sets the share of a flexible link, its share will be
a numerical value rather than "nil". This causes Link#update_share to
calculate the wrong value.

For example, if we have a slot of three links -- two type=share and one
type=flexible -- the share values pre-calculation might look like this:

    [0.25, 0.25, nil]

From this update_share would calculate the share of the flexible link:

    1.0 - 0.25 - 0.25 = 0.5

However if an input sets the share attribute of the flexible link, the
share values would be:

    [0.25, 0.25, 0.5]

... causing ETEngine to calculate the wrong share when updating:

    1.0 - 0.25 - 0.25 - 0.5 = 0.0

Now, we exclude the flexible link's own value when calculating a new
share value.

In practice, this was never really an issue since it only occurs when
there is zero demand through the slot (so would never result in
incorrect energy flows) and only if a share is manually set with an
input (therefore could not result in incorrect start values for inputs).