Prevent P2H electricity being used twice in MO
P2H being classed as having electricity input meant that the heat added
to P2H was being demanded as electricity in the next frame. This meant
that P2H was using twice as much electricity as it should:

  * as excess electricity in frame 0
  * as mandatory electricity demand in frame 1
  * etc...

This fixes P2H increasing blackout hours, when in fact this metric is
expected to remain constant or fall when P2H is installed.

Ref a296564