Fix demand subtraction in reconciliation subordinates
Restricts the leader to be a node with an appropriate output curve,
instead of supporting both inputs and outputs. By using the output curve
we can be sure that the demand subtracted is correct; using an input
curve may not correlate with output as energy may be stored in a buffer
rather than output immediately to satisfy demand.

Closes #1075