Dump unused stored energy in StorageAdapter
When a merit flexibility option stored energy which is not used in the
merit order calculation, the graph wrongly assumes that that energy
should still flow to the next node.

Now if the node has a share and an inversed_flexible links, it will
dump the unused energy through the inversed flexible.