Add support for rotating merit order curves
Curves used as inputs to the merit order may be optionally rotated so
that the calcualtion begins on a datetime other than Jan 1st 00:00. For
example, setting the start hour to 2190 (8760 / 4) begins the
calcualtion on April 1st 00:00, and runs to March 31st 23:00.

Curves are de-rotated before being set on the converter so that curves
in queries can be relied on to always begin Jan 1st.