Interleave heat network between Fever and electricity
Instead of calculating the heat network after Fever and the electricity
merit order, the heat network merit order is instead calculated hourly
between Fever and electricity.

This allows for dispatchable heat loads to create an electricity load in
the same hour. Previously it was not possible for the electricity load
to be accurately represented.