Enable heat storage for all protected scenarios
This sets the heat storage toggle to 'On' for all protected scenario's. The reason to do this is to reproduce the original energy flows of these scenario's more closely after the implementation of time-resolved heat networks. Enabling storage prevents heat produced during low demand hours from going to waste. This resembles the 'old' implementation of heat networks in which supply and demand are matched on a yearly basis.