Add support for waste_outputs in cost calculations
fuel_costs and co2_emissions_costs_per_typical_input will now compensate
for any outputs which are declared as loss. This is used in converters
where an inputs is converted into two or more carriers, but one carrier
is a waste byproduct, and should not be considered as part of the fuel
or CO2 cost.