Add Slot::LinkBased
Allows a slot to recalculate its conversion once the value of all links
are known. This helps when a Causality component changes the demand of
one or more links during the calculation.

This is typically the case when the heat network selects different
technologies during calculation, causing different demand for
electricity than was expected when the merit order was initialized.
In this case, using LinkBased ensures that the share of electricity and
loss is consistent with what actually happened during calculation.