Run Gquery behavior in the API
Rather than running the Gquery behavior in QueryInterface::Base, run the
behavior only when the data is being prepared to be sent to the user.

This prevents unnecessarily processing values when they are only used as
part of another query, and prevents CurveBehavior from messing up the
result of `SUM(Q(some_curve))` when the curve contains very small
values.