Fix input of carriers for HHPAdapter
Fixes the HHPAdapter so that the input carrier conversions are set in
the graph consistent with the values calculated by Fever. This means
that the shares of the `based_on` and `balanced_with` shares are set as
before, but the share of the secondary participant carrier (typically
network gas or hydrogen) is adjusted to reflect their output efficiency.

The output conversion of the converter is no longer changed, as these
changes would have knock-on effects on the flow of the primary carriers
(e.g. electricity and ambient heat), resulting in the electricity curve
and `input_of_electricity` no longer agreeing.

Closes quintel/etmodel#3223