MeritFacade: Prevent double-setting demand
Prevent a "follower" adapter from setting the demand of a converter when
another adapter has already done so.

For example, a CHP which participates in the electricity and heat
network merit orders should have the demand set by the electricity
component and not by the heat component. This is because the heat
component lags behind the electric by an hour, and therefore would
install a slightly incorrect demand and FLH.

Closes quintel/etmodel#3227