Add MeritFacade::CurtailedAlwaysOnAdapter
Enables curtailing the demand profile of always-on participants in
Merit. If the production_curtailment value is non-zero, the profile will
be saturated at a percentage of its peak value.

For example, a profile with a peak of 100 with curtailment set to 0.2
(20%) will cap all values at 80.

Converters whose values are curtailed will also receive a
`curtailment_output_curve` describing the energy from the original curve
which was removed.