Fix incorrect heat curves based on elec. flexibles
Fixes that flexibility options use their load curve to represent input
and output (input as negatives, output as positives). This would result
in incorrect curves when these were used as the basis for another
Causality component (e.g. when a heat network producer has a curve based
on the electricity input).