Support re-calculation of molecule graph
The molecule graph may now be used to determine some electricity
demands, to be used in Merit, and will then be recalculated after the
second energy graph calculation.

Note that energy demands set as a result of the molecule calculation
will not "flow" through the graph until the recalculation occurs; this
means that electricity (for example) demands must be set on nodes used
by the merit order so that these loads are accounted for.