Add support for `capacity_profile` user curve
These curves are expected to describe the hourly capacity use of the
technology, with values between 0 and 1. The FullLoadHours reducer has
been adjusted to reflect this, summing the values of the curve, rather
than being based on a normalized curve.

Using FullLoadHours with any other curve type will raise an error.