Move main Reconciliation setup after Merit/Fever
I moved these to earlier in the Causality calculation due to
quintel/etmodel#3605, but failed to account for Merit/Fever changing
the demand of Reconciliation nodes. This caused their hourly curves to
be created with pre-Merit/Fever demands.

Added a separate setup step for Reconciliation for adapters which need
pre-Merit demands, while moving the main setup to be after merit.

Closes quintel/etmodel#3649