Fix recursive factor with efficiencies exceeding 1
When the outputs of a node sum to a value greater than 1, recursive
factor values are exaggerated. For example, if we have a graph like
where a node outputs eneryg to two suppliers in different carriers, we
expect calculations on either of the inputs to be 50% of the input:

               [Supplier = 200]
                  /        \
    [Consumer 1 = 100]  [Consumer 2 = 100]

However, when the carrier shares exceed 1.0, calculations for some
recursive factor methods would be incorrect. For example, if both
outputs have a share of 1.0:

               [Supplier = 200]
                  /        \
    [Consumer 1 = 200]  [Consumer 2 = 200]

This is correct when thinking about demand, but not for primary demand.
For PD, each output is responsible for 50% of the input. The primary
demands should look like the first diagram.

This commit adds an `output_efficiency_compensation_factor` which is
used by `recursive_factor` (but not `recursive_factor_without_losses`)
to compensate for nodes with output shares exceeding 1.

The "without losses" does not have this compensation, as it is typically
used to calculate shares (`sustainability_share`).

Fixes #1172