Ensure electrolyser shares are consistent
Sets electrolyser shares prior to recalculation of the graph. 2d5733f
required that Reconciliation be moved to after recalculation of the
graph so that demand changes made by Merit/Fever were reflected in
hourly network gas and hydrogen.

This had the effect of setting electrolyser shares too late: the graph
would already have been calculated, and flows were therefore not
consistent with the shares set by ElectrolyserAdapter.

As electrolyser demand is known prior to Merit, and cannot be changed
(due to them having dedicated electricity sources which are excluded
from Merit), we can set their demands prior to recalculation of the
graph.

A `before_graph_recalculation` method is added to Reconciliation which
is triggered by Causality.

Closes #1174