Fix sustainability_share with non-unity inputs
When sustainability_share was called on nodes where the path to primary
demand has non-unity inputs, the share would be incorrectly calculated.

    +----------+ - - - - - +      + - - - - - +----------+
    | Consumer | gas @ 0.6 |  <-  | gas @ 1.0 | Producer |
    +----------+ - - - - - +      + - - - - - +----------+

If the producer has a `sustainability_share` of 0.5, then so should the
consumer, since the producer is the sole supplier of energy. However,
without compensating for the non-unity input the share would be
multiplied by 0.6, resulting in the wrong recursive value on the
consumer of 0.5 * 0.6 = 0.3.

`recursive_factor_without_losses` now supports a `value_type: :value`
option, ensuring that such shares are compensated for when propagating a
value from right-to-left, rather than a factor.

Closes #1181