Fix mismatch between heat curves and output
When a heat producer also participates in the electricity merit order,
the heat curve does not have a load to represent the final hour in the
year (all heat loads a shifted one hour into the future, so this
manifests with no heat load in the _first_ element of the load curve).

This results in slightly too much heat leaving the producer in the graph
since the share of heat output does not account for the hour in which
none is produced.

ProducerAdapter has been adjusted to change the share of heat (and loss,
if applicable) to ensure that graph flows match the heat output curve.

Closes #1175