Use Fever config to calculate CoP when possible
When a Fever config is present on a node, and that config defines a
"based on" and "balanced with" carrier, calculate the CoP using these
carriers ignoring any other inputs to the node.

This fixes the CoP for hybrid heat pumps which should not include inputs
from the secondary (typically gas or H2) component.

Closes #1185