Add support for always-on battery parks
These parks consist of an always-on producer connected to the HV network
with a capacity-constrained interconnect. When the energy produced
exceeds that of the interconnect energy is instead stored in a battery.

Energy is discharged from the battery whenever the output of the
producer is lower than that of the interconnect.