Add optimized storage flexibility technology
Computes a residual load curve (based on inflexible supplies and
demands) The battery will charge and discharge in a manner than flatten
the curve as much as possible.

Within Merit, these batteries are represented as two separate
technologies:

* A user which consumes energy.
* A consumer which produces energy.

Consumption and production are determined ahead of Merit by the
StorageAlgorithm module.