Add `capacity_costs` to node
Adds a `capacity_costs` method which calculates a fixed cost for the
node based in the input capacity. In most cases this is expected to be
zero, but in some cases a `fixed_costs_per_mw_input_capacity` value will
be defined, and an additional fixed cost is added to the node based on
the input capacity.

Bumps Atlas to quintel/atlas@afabbe7

Ref #1207