Prevent protected scenarios from being updated
Uses cancancan to authorize requests to read or update scenarios.
Presently all scenarios are world readable and writeable. This commit
changes protected scenarios, making them immutable once protected is set
to true.

This also means that the original creator of the scenario will not be
able to make further changes to the scenario, but will be able to create
clones which can be edited further.