Fix heat network following price-sensitive flex
We (ab)use the marginal_costs attribute for heat network participants to
sort them according to the user's choice. However, the first
participant was assigned a cost (willingness-to-pay) of 0, which meant
that it would never receive energy.

This commit also disables consumption from dispatchables, so that only
excess is used by flex options.