Add CodeMirror, revamp scenario editor
Adds CodeMirror to syntax highlight the GQL edit form, as well as user
and balanced values on scenario forms, and metadata.

Adds balanced values and metadata fields to the scenario editor.

Removes the old "_as_yaml" methods and extracted updates for scenarios
through the inspect interface to a new `Scenario::Editable` helper
class.