bin/import-db should create but not set up DBs
Running db:setup would create the DB tables. This causes an issue when a
migration not present on the imported file would create a table which is
already created by db:setup.

db:create will create the databases without tables, which can then be
imported through the SQL file and migrated as normal.