Set user values for vans when trucks inputs were touched
Part of the splitting up of trucks and vans into separate applications.
Copies the technology splits, and recalculates the application splits for trucks and vans based on the dataset's default share for trucks. These default shares are read from a csv, as we don't want to recalculate this value for each scenario.