Skip availability check when a curve is present
When a merit order participant has an availablility curve attached, we
should not exclude it from the calculation just because the user
previously set its numeric availability to zero.

Instead, we consider that having an attached curve means that it has
non-zero availability and can be included in the calculation if all
other criteria are met.

Closes #1260