Exclude curve-disabled inputs from scenario
Do not run inputs which are disabled by an attached curve. Previously,
the value would still be set with only the front-end sliders being
disabled. This meant that a value set by a user may still have an effect
on the scenario even though the slider was disabled in the front-end.