Allow inputs API to return original defaults
The inputs API returns data about all the inputs for a scenario,
including the minimum and maximum permitted values, as well as the
default.

When a scenario has a parent, rather than returning the normal defaults,
we'll return the values from the parent dataset instead. This allows the
"reset" button in the ETM to return a child scenario to the values from
the parent, rather than the original default values.

This commit adds the ability to send a `defaults` query parameter set to
"original". This overrides the behaviour and will result in the API
always sending the original dataset defaults, instead of those from the
parent.

Ref multi-year-charts#18