Add an API for saved scenarios
Adds a controller for listing, getting, creating, updating, and deleting
saved scenarios. The controller receives the user request, forwards it
to ETEngine with a JWT token with the same scope as the user provided,
and then responds back to the user.

Saved scenario information is decorated with informtion about the
scenario. It is not possible to make changes to the scenario itself;
users should get the `scenario_id` from this API and then perform their
desired actions on the scenario.